refactor: refactor the stop log server function.
